Make TypoMiner non-template
TypoMiner itself is not required to be a class template, none of its
methods use template arguments. Furthermore it is not possible to
create an object of type TypoMiner without specifying template arguments
making it unusable.
Implements static template method to create an instance of TypoMiner,
moves all method definitions to .cpp file.